backend/pyproject.toml
[build-system]
requires = ["uv_build>=0.10.10,<0.11.0"]
build-backend = "uv_build"

[project]
name = "m00d-backend"
version = "1.0.0"
description = "Mood tracker (Backend)"
license = "GPL-3.0-only"
authors = [
    {name = "Chris Post", email = "admin@postfmly.com"}
]
keywords = ["mood", "tracker"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment :: Mozilla",
    "Framework :: FastAPI",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: English",
    "Operating System :: POSIX :: Linux",
    "Private :: Do Not Upload",
    "Programming Language :: Python :: 3.14",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
    "Typing :: Typed"
]
dependencies = [
    "fastapi[standard-no-fastapi-cloud-cli]>=0.135.1",
    "peewee>=4.0.2",
    "pydantic>=2.12.5",
    "python-box~=7.0",
    "rich>=14.3.3",
]
requires-python = ">=3.14"

[project.urls]
GitHub = "https://github.com/chump29/m00d"

[project.optional-dependencies]
dev = [
    "behave>=1.3.3",
    "pip-review>=1.3.0",
    "pylint>=4.0.5",
    "types-peewee>=4.0.0.20260223"
]

[tool.pylint.master]
ignore-paths = [
    ".venv"
]

[tool.uv]
python-downloads = "manual"

[tool.uv.build-backend]
namespace = true
source-exclude = [
    ".dockerignore",
    ".python-version",
    "*.sh",
    "/features",
    "Dockerfile",
    "README.md",
    "uv.lock"
]
